test: address review feedback on thread reply regression test
- Move test to separate thread_reply_tests.rs file (app.rs tests ~1700 lines)
- Fix vacuous assertion: write reply2 after cursor-at-EOF, then refresh()
  exercises the live-message routing path and verifies non-accumulation
- Fix doc: reference draw_chat_messages() not visible_messages() as bug site
- Remove unnecessary 50ms sleep (channel.send() releases lock synchronously)
- Add test_app_with_channel() helper to app.rs for external test files

🌃 Co-built with [Midtown](https://github.com/btucker/midtown)
